+++ .gitignore
@@ -3,3 +3,5 @@ node_modules/
 screenshots/
 oldscreens/
 .env
+credentials.json
+token.json
