+++ lib/ophanHelper.js
@@ -31,7 +31,7 @@ const tags = [
     "tone/recipes",
   ];
 
-const amountPerTag = 10; // configurable option?
+const amountPerTag = 10; // make configurable option?
 
 const ophanAPIUrl = tag => 
     `https://api.ophan.co.uk/api/mostread/keywordtag/${encodeURIComponent(tag)}?count=${amountPerTag}`
