+++ package.json
@@ -10,11 +10,12 @@
     "axios": "^0.19.2",
     "commander": "^8.1.0",
     "dotenv": "^10.0.0",
-    "google-spreadsheet": "^3.1.15",
+    "googleapis": "39",
     "merge-img": "^2.1.3",
     "node-fetch": "^2.6.0",
     "puppeteer": "^1.18.1",
     "puppeteer-cluster": "^0.16.0",
+    "readline": "^1.3.0",
     "take-screenshots": "^0.1.0"
   }
 }
